refactor: move `CommandDisplay` to molecules
