+++ src/components/molecules/CommandDisplay.tsx
